Weighted random drive choice made
every cycle the director measures all the drives in the game and records
which drive is most prevalent - ie, is the conversation mostly angry,
envious, lustful etc.

The director uses this information to weight the newly created gambits
towards increasing whatever is prevalent in the conversation.

This should create a positive feedback driving whatever is the most
prevalent drive higher - an angry conversation should get more angry for
example.